Moved Timer Logic to PossessableObject script
