+++ Geist Heist/Assets/Scripts/Player/Possession/PossessableObject.cs
@@ -11,15 +11,38 @@
 using UnityEngine;
 using Unity.Cinemachine;
 using NaughtyAttributes;
+using UnityEngine.UI;
+using System;
 
 public class PossessableObject : MonoBehaviour, IInteractable
 {
     public IInputHandler InputHandler;
     [Required] public CinemachineCamera CinemachineCamera;
+    [Header("Timer Variables")]
+    [SerializeField] private bool hasTimer;
+    [SerializeField] private float timerTime = 5f;
+    private float currentTimerTime;
+    private bool isTimerActive;
+    [SerializeField, Required] Slider timerSlider;
 
     private void Awake()
     {
         InputHandler = InputHandler ?? GetComponent<IInputHandler>();
+    }
+
+    private void Update()
+    {
+        if (isTimerActive && hasTimer)
+        {
+            currentTimerTime -= Time.deltaTime;
+            if (currentTimerTime <= 0)
+            {
+                currentTimerTime = 0;
+                isTimerActive = false;
+                OnTimerFinished();
+            }
+            UpdateSlider();
+        }
     }
 
     void IInteractable.Interact()
@@ -31,11 +54,45 @@ public class PossessableObject : MonoBehaviour, IInteractable
     public void OnPossessionStarted()
     {
         InputHandler.OnPossessionStarted();
+        if(timerSlider != null)
+        {
+            timerSlider.gameObject.SetActive(true);
+        }
+        isTimerActive = true;
+        currentTimerTime = timerTime;
     }
 
     // Called in PlayerManager
     public void OnPossessionEnded()
     {
         InputHandler.OnPossessionEnded();
+        if(timerSlider != null)
+        {
+            timerSlider.gameObject.SetActive(false);
+            ResetTimer();
+        }
     }
+
+    #region Timer
+    private void OnTimerFinished()
+    {
+        PlayerManager.Instance.PossessGhost(gameObject.transform.GetComponent<PossessableObject>());
+        ResetTimer();
+    }
+
+    private void UpdateSlider()
+    {
+        if (timerSlider != null)
+        {
+            timerSlider.value = currentTimerTime / timerTime;
+        }
+    }
+
+    private void ResetTimer()
+    {
+        currentTimerTime = timerTime;
+        isTimerActive = false;
+        timerSlider.gameObject.SetActive(false);
+    }
+    #endregion
 }
