+++ Geist Heist/Assets/Scripts/Player/Possession/VendingObject.cs
@@ -23,46 +23,21 @@ public class VendingObject : IInputHandler, IInteractable
     /*[Dropdown("balancing")]*/[SerializeField] private Vector3 launchDirection;
     /*[Dropdown("balancing")]*/[SerializeField] private bool Tap;
 
-    [Header("Timer Variables")]
-    [SerializeField] private float timerTime = 5f;
-    private float currentTimerTime;
-    private bool isTimerActive;
-    [SerializeField] Slider timerSlider;
-
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
     {
         
     }
 
-    // Update is called once per frame
-    void Update()
-    {
-        if (isTimerActive)
-        {
-            currentTimerTime -= Time.deltaTime;
-            if (currentTimerTime <= 0)
-            {
-                currentTimerTime = 0;
-                isTimerActive = false;
-                OnTimerFinished();
-            }
-            UpdateSlider();
-        }
-    }
-
     public override void OnPossessionStarted()
     {
-        timerSlider.gameObject.SetActive(true);
-        isTimerActive = true;
-        currentTimerTime = timerTime;
+
     }
 
     public override void OnPossessionEnded()
     {
-        timerSlider.gameObject.SetActive(false);
-    }
 
+    }
 
     #region action
     public override void OnActionStarted()
@@ -116,7 +91,7 @@ public class VendingObject : IInputHandler, IInteractable
     { }
     public override void WhileActionNotHeld()
     {
-        throw new System.NotImplementedException();
+        //throw new System.NotImplementedException();
     }
 
     public override void OnInteractCanceled()
@@ -138,32 +113,8 @@ public class VendingObject : IInputHandler, IInteractable
 
     }
 
-
     public override void OnMoveCanceled() { }
 
-    #endregion
-
-    #region Timer
-    private void OnTimerFinished()
-    {
-        OnInteractStarted();
-        ResetTimer();
-    }
-
-    private void UpdateSlider()
-    {
-        if (timerSlider != null)
-        {
-            timerSlider.value = currentTimerTime / timerTime;
-        }
-    }
-
-    private void ResetTimer()
-    {
-        currentTimerTime = timerTime;
-        isTimerActive = false;
-        timerSlider.gameObject.SetActive(false);
-    }
     #endregion
 
     void IInteractable.Interact()
